// frontend/src/shared/api/generated/assessments/models/assessmentInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Assessments
 * OpenAPI spec version: 0.1.0
 */
import type { Grade } from './grade';
import type { ITRole } from './iTRole';

export interface AssessmentInfo {
  id: string;
  slug: string;
  title: string;
  description: string;
  specialization: ITRole;
  grade: Grade;
  time_limit_seconds: number;
  tasks_per_attempt: number;
}