// frontend/src/shared/api/generated/assessments/models/surveyOptions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Assessments
 * OpenAPI spec version: 0.1.0
 */
import type { CooldownInfo } from './cooldownInfo';
import type { Option } from './option';
import type { SpecializationOption } from './specializationOption';
import type { SurveyPrefill } from './surveyPrefill';

export interface SurveyOptions {
  industries: Option[];
  specializations: SpecializationOption[];
  grades: Option[];
  prefill: SurveyPrefill | null;
  cooldowns: CooldownInfo[];
  pass_percent: number;
  exceed_percent: number;
}